Implement seekable readers as their own types
This solves the problem of the reader's stream initially being
positioned ahead of the start - like if we've had to skip over
a bunch of ID3v2 junk in order to get to the start of the FLAC data.

I could just have something like FlacReader::open_seekable()
if the input stream is, in fact, seekable, but users are likely
to forget that.  Having seperate types with more restrictive bounds
are less likely to be messed-up.