Add preliminary cue sheet parsing
Because CUESHEET blocks have a long litany of rules about
what's acceptable and what isn't, handling this stuff
takes a tremendous amount of effort for something that
amounts to a single class method attached to the Cuesheet struct.

But it does pass most of the reference implementation's
tests, at least.