Unify the cuesheet errors under a single type
Doesn't make a lot of sense to have a different
set of errors when reading a CUESHEET block from a stream
vs. parsing it from a file.  There are a few that
only occur during reading, and some that only occur during parsing,
but it's best to keep them all together in one place.