Add an option for fast channel correlation
I don't believe this is an option in the reference implementation.
The new default behavior is to do an exhaustive test by calculating
both into recorders and picking the best pair.  Whereas the fast
check simply calculates the abs sum of the samples
and chooses the smallest pair between those before encoding subframes.

The former approach tends to encode a bit slower, but also a bit better
by a few percentage points.  So that seems like a good candidate
for an encoder flag.