Bump version to 1.0.0
This has been a long time coming.

I'm sure there'll be plenty of additional changes down the line
and I'll need to dump up to bigger and better version numbers
in the future, but I'm generally happy with the shape of the
crate's API and won't be making big changes without very
good reasons.