Add more convenience methods for adding metadata blocks at encode-time
A new marker trait is used to accept only blocks which can
be safely moved from one FLAC encoding to another.

For example, the contents of the STREAMINFO and SEEKTABLE
blocks are calculated by the encoder and we don't want users
attempting to specify their own.  Even if attempting to re-encode
the same file, different encoding parameters may result
in different metadata blocks.